add event emitters to the updated ProfileService
